Review and fix scaffold issues.
This commit addresses findings from a comprehensive review of the project scaffold.

Key actions taken:
- Standardized `.gitignore`.
- Corrected structure of `src/main.py` and `src/__main__.py`.
- Implemented example `greet()` and `run_project_core_logic()` in `src/main.py`.
- Removed `print()` on import from `src/__init__.py`.
- Added basic error handling to `src/__main__.py`.
- Instantiated global `settings` in `config.py`.
- Ensured example tests in `tests/test_example.py` align with corrected code.
- Added `CODE_REVIEW_AND_FIXES.md` detailing all findings and actions.

The scaffold should now be more robust and provide a better starting point for new projects.
