+++ {{cookiecutter.project_slug}}/src/{{cookiecutter.project_slug}}/main.py
@@ -1,32 +1,27 @@
-# {{cookiecutter.project_slug}}/src/{{cookiecutter.project_slug}}/__main__.py
+# {{cookiecutter.project_slug}}/src/{{cookiecutter.project_slug}}/main.py
 """
-Executable module for the {{ cookiecutter.project_name }} package.
-
-This allows the package to be run using 'python -m {{ cookiecutter.project_slug }}'.
-It typically calls the main function or entry point of the application.
+Main module for {{ cookiecutter.project_name }}.
+This module can contain the primary business logic or core functions of the application.
 """
 
-from .main import run_project_core_logic
-# You might also import and parse command-line arguments here if not done in main.py
-# import argparse
+from {{ cookiecutter.project_slug }} import __version__
 
+def greet(name: str = "World") -> str:
+    """Returns a greeting string."""
+    return f"Hello, {name}, from {{ cookiecutter.project_name }}!"
 
-def main_cli():
+def run_project_core_logic():
     """
-    Command-line entry point.
-    Parses arguments (if any) and calls the main application logic.
+    Placeholder for the main execution logic of the project.
+    This could be expanded to orchestrate various tasks, initialize resources,
+    or start an application server if applicable.
     """
-    # Example: If you were to add argument parsing:
-    # parser = argparse.ArgumentParser(description="{{ cookiecutter.description }}")
-    # parser.add_argument("--config", help="Path to a configuration file.")
-    # args = parser.parse_args()
-    #
-    # run_project_core_logic(config_path=args.config)
+    print(greet()) # Example: Call the greet function
+    print("Project core logic is running.")
+    # TODO: Implement actual core logic here.
 
-    # For now, we'll directly call the main logic function.
-    print(f"Executing {{ cookiecutter.project_name }} via __main__.py...")
+if __name__ == '__main__':
+    # This block allows running this module directly for testing or specific tasks,
+    # but it's not the primary entry point if __main__.py is used.
+    print("Executing main.py directly (for testing/dev purposes)...")
     run_project_core_logic()
-
-
-if __name__ == "__main__":
-    main_cli()
